refactor: :recycle: refactor: Refatoração de código no servidor
